+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 All notable changes to the ESP-IDF Web Extension for VS Code will be documented in this file.
 
+## [0.0.5]
+- Use `idfWeb.buildPath` for flash and monitor artifacts (empty default is `${workspaceFolder}/build`)
+- Support CH340 with WebUSB when `idfWeb.useWebUsbCh340` is set to true.
+- Add IDF Monitor Colors
+- Allow the `ESP-IDF-Web Monitor` to run without a workspace folder and add a `idfWeb.monitorBaudRate` for fallback if no project_description.json or workspace folder is found.
+- Update esptool-js to v0.7.0 to support ESP32-S31 and newer chips that use GET_SECURITY_INFO to detect chips and fix monitor.
+
+
 ## [0.0.4]
 - [Fix device disconnect callback](https://github.com/espressif/vscode-esp-idf-web-extension/pull/18) Thanks @archef2000
 - [Add monitor input functionality use native api string conversion](https://github.com/espressif/vscode-esp-idf-web-extension/pull/19) Thanks @archef2000